Always start/stop Range background goroutines via the Store.
Rename Store.Init and Close to Start and Stop to match other
methods that manage background goroutines.

This change ensures that we are always calling Store.Start
(and hopefully always calling Store.Stop, although I haven't
verified that we have 100% coverage), in preparation for
adding a new Store-level background goroutine.