Call, Prepare/Flush, and RunTransaction examples.
Example functions for a simple Put & Get using Call(), Prepare() &
Flush(), and then RunTransaction().

Each example is self-contained, and includes the server
setup/teardown.  This makes the code closer to what would be used in
an actual client program.

Named all examples so that they will appear in the correct section
of the documentation.

The Prepare() & Flush() is a bit long because we have to create the values
we are later scanning for.

All the examples have been updated to match the coding style and
fix other issues.  Prepare/Flush example now runs multiple scans for
sub-ranges of the keys in parallel.