Upgraded the gossip callback to include contentsChanged bool.
This informs the callback callee whether or not the actual contents
of the info were updated. This can save a lot of work in the event
that the new info just has a shorter number of hops or an updated
timestamp but the same contents--we expect this to be a common
occurence.