Remove all instances in test code which use a zero timestamp.
This is in preparation of using a zero timestamp to mean an inline
MVCC value so we can avoid ever using the engine directly and instead
write all data through the MVCC interface, including merged stat
counters, txn entries, response cache entries, time series data,
and other system-local data.

After some thought, Matt and I ruled out the possibility of not
inlining the value and just using a single version--not much more
efficient and would allow us to use the MVCC code exactly as is.
However, this would have made merging difficult, as there would be
no way to create the MVCC metadata on the first merge.

Support for both types of values is becoming increasingly unwieldy
and the prospect of a parallel interface for querying time series
data and the bifurcation of different regions of the map to be
"mvcc" and "non-mvcc" seems like a pandora's box of troubles.

Follow-on commits in this branch will implement the actual inline
value in proto.MVCCMetadata and the corresponding improvements to
mvcc.go. I intend to take this opportunity to substantially improve
the performance of mvcc scan.