Add InternalMerge Command
Add the InternalMerge command, which is replacing the anticipated (but never
implemented) AccumulateTS command.  InternalMerge is similar to a put in that it
writes a single value to a key - however, the supplied value is *merged* into
the existing value instead of replacing it.

Merging is not transactional, and thus it is only exposed internally for usage
in certain operations which require a high write performance (such as the
accumulation of time series data).