GC stats to compute likely GC'able bytes from elapsed time.
A GCMetadata record will be written after every GC scan of a range. That,
combined with the current non-live bytes count and current time can be
used to estimate the expected count of GC'able bytes.

This will allow us to avoid constantly scanning ranges which have had no
churn, while using the range stats as well as a detailed picture of the
aging bytes on last GC scan to prioritize ranges for GC.