Log instead of panicking when ProposeRaftCommand called after shutdown.
When a test that uses goroutines fails, these goroutines may continue
to try to use the store after it is shut down. The panic occurs during a
subsequent test function and obscures the true cause of the failure.

This problem appears to occur most often with TestRangeIdempotence,
so I've increased its timeout.

Fixes #222.