Add basic KV REST html UI and embedfile tool
This change adds a basic UI for querying the REST API from a web GUI at /kv/rest/.
It avoids the external dependency of another file (rest_ui.html) by generating a
file through the go generate command and the embedfile tool before compilation.

This is all a preamble to a better REST API, but also autogenerated documentation
using embedfile and go generate.