Reimplement snapshots as a new implementation of Engine interface.
This fixes a TODO in the code and helpful to upcoming changes to
implement asynchronous scans of ranges for custodial work. Without,
more Engine method variants taking snapshot IDs would be necessary.
This change is more in keeping with how we do Batch engines and is
simpler.

Also remove the InternalSnapshotCopy method, which will be replaced
by an in-Raft mechanism whereby the leader feeds the snapshot directly
to a requesting follower.