Scan queue for range maintenance.
Implementation of a range data iterator which scan through all of the
data in a range. This is now used for Range.Destroy(). It will also be
used for the range snapshot when instantiating a new range Raft replica.
Here, it's used to scan over the complete set of data in the range.
Versions for the same key will be processed for garbage collection.

Changed transaction key to encode base key to guarantee that all txn
records in a range sort together.

Implementation and unittests for scan queue's ShouldQueue method,
which combines priorities from GC'able bytes, last time since
unresolved intents were scanned (if there are any), and last time
since the full range had its checksums verified.

Next up will be implemention of scan queue's Process method, which
will implement garbage collection and intent resolution.