Revert to processing batched requests serially.
The previous code would sometimes reorder individual batch calls,
which occasionally caused the EndTransaction to be processed before
other ops in the batch. This change reverts to a serial execution.

The code wasn't correct in any case as we really need to have the
distributed sender handle batch requests in order to separate batches
into those which may be executed on the same range. I've left this as
a TODO.