Reorganized all local keys into two sections/range: by ID & by Key.
This groups all of the range-local data contiguously into two sections,
making range traversals require less seeks and more importantly, less
maintenance as new range-local data is added going forward.

The trickiest part of this was upgrading the C++ code to properly
account for encoded keys, since it now has to decode the key and skip
the range ID or range Key to determine whether the tuple being compacted
is a transaction record or response cache entry.