Replace use of pkg-config with environment variables in Makefile.
Some go tools (such as go oracle and gorename) do not understand files
that use the pkg-config feature or even depend on it indirectly. With
this change, such tools can be used, although in exchange the
CGO_{CPP,LD}FLAGS environment variables must be set. This is done
automatically in the Makefile (along with the GOPATH) but it must be set
separately to use 'go test' directly.