tag Docker images and publish cockroachdb/cockroach-dev
the cockroachdb/cockroach-dev image was not built automatically any more
because we introduced the Dockerfile template (and hence there was no
vanilla Dockerfile available for the registry to build).
Instead, we now push that image from the CircleCI release process.

Additionally, the cockroachdb/cockroach{,dev} images are now tagged
with the output of `git describe`.

I have pushed an `alpha` tag pointing to the initial commit in this
repo so that `git describe` works properly.