Refactor flags into a central package.
Moving flags has the benefit of more inspectable and true package interfaces and allows embedding the core server process into other systems or experiments. The downside is increased plumbing of values. Fortunately this additional plumbing is mostly minimal. The API for the package "server" has increased, but only what is needed for the cli package.

Future work may remove or refactor the TestServer to directly use the server.Context struct or use the server.Server directly.

The "server/cli" package has been added and rightly might belong in the main package. But for the sake of keeping the main package small, this has been kept separate.