checkout & avoid changing timestamps on submodules
in #309, we realized that there was a weird dependency problem where
the docker images ended up with the wrong version of RocksDB baked in,
despite earlier efforts to bake the SHA into the lower layers.

this commit fixes a possible reason for this, namely the fact that submodules
may not have been updated appropriately when the git repo was cached by
CircleCI.

prior to this change, all files' timestamps were reset and then updated
according to their last commit. this did not apply to submodules though,
so what could have happened is that a submodule containing an earlier build
would be forwarded, and the build files' timestamp reset.

then, rocksdb may have been built from the correct version, but old build files
with reset timestamps may have confused make into building a broken version
of rocksdb.

i suspect that this fixes #309. we have since removed rocksdb here, but this
change will preserve this from happening again with other etcd or others.