Replace the InternalChangeReplicas raft command with a distributed txn.
This allows us to update the addressing records as well, but I'm leaving
that for a subsequent commit.

multiraft.ChangeGroupMembership can now smuggle a command to be executed
atomically with the membership change; we do this to attach the
membership change to the EndTransaction call.

storage/raft_test.go has been moved to client_test.go and
client_raft_test.go now that it needs to use distributed transactions.