Split ScanQueue into GCQueue and VerifyQueue.
ScanQueue used to try to treat GC and verification scan in the same
pass, but this only works for the leader. The GC scan, which handles
old MVCC values and finds the oldest extant intent age, is only done
on the leader replica. However, verification must be done independently
on each replica, as checksum failures can of course occur on all replicas.

Added processing functionality to baseQueue. This starts up a goroutine
per queue to process entries and ties the speed of processing to a new
timer function which is passed during queue creation. In this manner,
verification of range on-disk checksums can be spaced over the course
of the entire verification interval (current 60 days), so the system
doesn't try to greedily process verifications.