Fix a race in restoring from on-disk state.
Don't start the raft threads until we have loaded all the ranges.
Raft was sometimes starting to operate on ranges that hadn't been
initialized yet.

Lazily initialize the consensus groups for each range. This was
a necessary side effect of this change but also addresses the
election storm on cold start as discussed in #357.