Update etcd submodule.
This revision is the upstream master branch; our fork is now obsolete.
The .gitmodules file still points to cockroachdb/etcd.git because
it is not possible to change a submodule url without manual intervention
by all developers, and the submodule is going away soon anyway.

Update multiraft to use ReportUnreachable and ReportSnapshot
(the latter is now mandatory).

Fix multiraft/storagetest which was broken by upstream changes to
MemoryStorage.