embed test_certs, remove reliance of tests on cwd
the test certs are now embedded in package resource.
refactored the tls testing code so that it just pulls the asset.

as a nice side effect, the `make testbuild` task got easier since it doesn't
have to move the tests binaries to the correct location any more. the deploy
image creation script simplifies accordingly.