auto-open issues for failed tests
unfortunately, CircleCI as of now has no way of checking
via circle.yml what the test status is (unless we add a
hack for each test, which is clumsy), so we grep the
log files for interesting info and open issues whenever
something that looks like a test failure is found.

using issues, it will be easier to assign to fixing a par-
ticular build.