remove the unnecessary loop condition "exit"
After the retry logic is added, no need to set "exit" condition when receiving error. After all,  there will not be a condition like: "ids is empty" and "retry times is more than maximum",  and at the same time, "exit" flag is false. In the previous logic, if retry times increases, "exit" will always be set to true.