Refactor error protos.
Move the error union to ErrorDetail and move GenericError fields to
the top-level Error proto, and set them even when a structured detail
is available. This ensures that basic error information can always
be printed, even in non-go clients that don't have the methods
defined in proto/errors.go, or when there is version skew between
client and server.