Fix deadlock in TestRangeSplitsWithWritePressure.
The shutdown sequence is delicate; we need to stop everything that might
generate new work before stopping the parts that are necessary to
complete that work. I think that ultimately we'll need to have multiple
stop-like methods for different phases of the shutdown process but for
now this appears to eliminate the deadlocks we've been seeing recently.

Fixes #480.