Set Header.RaftID in RPCs sent by DistSender.
This speeds up handling of RPCs on the server since the range can be
looked up by ID instead of by key. It also allows DistSender to
determine when its range descriptor cache is out of date since the
recipient of the rpc will return an error causing the cache entry to be
evicted if the range does not exist on the node.