Directly resolve intents for keys in special circumstances.
Using the InternalCommitTrigger option inside of EndTransaction, keys
which are known to fall within the same range as the transaction may
be preemptively resolved at commit time. This prevents the possibility
of any intents existing on certain keys after their transaction has been
committed. This is important for the case of transactions which modify
range descriptors, including splits, merges & replica changes.

When the range descriptors are scanned at startup and when read during
raft snapshot replays, dangling intents are very problematic, if there's
a chance that the transaction which created them might have already been
committed. This change prevents that.