Fix a small bug in parseBootstrapAddresses()
len(strings.Spllit("", ",")) returns 1, so the previous 'if' condition
for no bootstrap node returns false even if g.gossipBootstrap is an
empty string.

When I changed the check to g.gossipBootstrap == "", several tests got a panic in the following case:

1) SetBootstrap([]net.Addr{s.rpc.Addr()}) is called for self-bootstrap
   and the node's address is added to g.bootstraps.
2) bootstrap() calls parseBootstrapAddresses() and removes the its own
   node address from g.bootstraps. g.bootstraps becomes empty.
3) g.Gossip() is called and parseBootstrapAddresses() is called again
   from bootstrap(). The check fails thsi time since g.bootstraps is empty.

Working around this by adding '!g.isBootstrap' to the 'if' condition.