Fix bug in clientCodec.WriteRequest and serverCodec.WriteResponse.
WriteRequest and WriteResponse need to be safe for concurrent use. Both
were writing out the header and body of a request to the underlying conn
without proper locking which would allow interleaving of the data.

Discovered this as part of some network optimization work which is also
present in this change. Wrap bufio readers and writers around the
connection. Use the optimized marshalers and unmarshalers for the
request and response headers. Remove snappy compression which benchmarks
as a pure loss (presumably due to the Go implementation). Removed
checksumming of the body which isn't present in the other Go rpc codecs
and doesn't seem worthwhile given we have end-to-end checksums of user
data along with the TCP checksums. Reduced allocations in various
network code paths.

benchmark                      old ns/op     new ns/op     delta
BenchmarkEchoGobRPC            45062         43120         -4.31%
BenchmarkEchoProtobufRPC       101950        38029         -62.70%

benchmark                      old MB/s     new MB/s     speedup
BenchmarkEchoGobRPC            22.72        23.75        1.05x
BenchmarkEchoProtobufRPC       10.04        26.93        2.68x