use leader cache and attribute lists
... to determine where to send RPCs.

If the call needs a leader and we think we know who that is,
send the call there first.

If we have attributes for our node and some of the replicas
match parts of those attributes (in a way that suggests geo
graphic proximity), prefer those.

Remark: The leader cache is not filled yet, but will be soon.