Make shutdown more robust.
We had whitelisted the txnMetadata.close goroutines in leaktest, but
the fact that they were still running after shutdown would sometimes
leave other threads to be leaked, or cause nil pointer panics.

This commit adds additional checks in retry loops that may be running
after shutdown to end the loop cleanly and allow all goroutines to
exit. This lets us remove the whitelisting of txnMetadata.close.

Closes #597