Add ts Package, Time Series Key Encoding
This commit creates the /ts package for handling time series data.

Included in the initial commit is the encoding and decoding logic for time
series data keys. This logic constructs cockroach keys based on the series name
and timestamp for datapoints; this should providing a usefully ordered
distribution of time series data.