Add a test case where offset errors are bigger than the max offset
This kind of scenarios could happen when heartbeat messages to the
node are having high latency. The computed cluster offset interval
becomes a wide range, and it might be considered healthy even when
the clock offset is bigger than maxOffset.

Not sure it's worht adding this test, but it was useful for me to
understand how the system works..