address @spencerkimball's feedback
most notably, make sure that the rpc server always returns an address that we
hope other clients can use to connect to us. of course, this means that the
hostname must be resolvable by those clients.
i'm assuming that we'll have to revisit the issue of having to know the
server's identity to pass to the outside world again when dealing with TLS,
and maybe we'll then decide that listening on all interfaces isn't even worth
it. in any case, for now it should work as advertised above.