Introduce util.SucceedsWithin, a replacement for IsTrueWithin.
Since this is a test-only function, it now takes the testing.T
and calls t.Fatalf directly, saving some syntactically-awkward
boilerplate at each call site. Returning an error instead of a bool
also allows us to produce better error messages in case of failure.
Finally, I swapped the order of the arguments because I think it
is aesthetically nicer for the inline function to be the last
argument.