Add Feed messages to Storage package
This commit creates a set of store-specific event messages meant to be published
over a feed. These are produced by the `StoreEventFeed` class, which provides
helper methods for publishing each available message type.

Also included is the `newBatch()` method, which returns an object used for
batching store methods. This was created to support some of the more complicated
range processes, where the data needed for an event is available significantly
before the process which produces the event is actually committed.

Also includes a `Difference` method for proto.MVCCStats.