Simplify Feed System
While investigating the addition of feeds to the Store system, two requirements
came under scrutiny:
1. We must be able to subscribe to a feed before it 'starts'.
2. It is NOT necessary for StoreEventFeed to maintain commitable batches.

The second requirement was repaired by removing StoreEventFeed's newBatch method
and associated structures, a significant simplification.

The first requirement was solved by converting feeds to a simpler "Big Channel"
model, where each Subscriber maintains a very large buffer for its channel of
events. This is a dramatic simplification; Feeds no longer need to be started,
and no additional goroutines are created to service them.

While it's possible that this new setup could panic if a Subscriber
doesn't read its channel, this kind of performance should not be a concern at
the frequency we are intending to publish events.