Publish Store events to Feed
Stores can now publish interesting events to a util.Feed. The target feed is
passed to NewStore as part of the StoreContext. Passing a feed is optional; if
no feed is given to a store, then the store will not publish events.

This commit also augments MultiTestContext to optionally pass a feed to the
Stores it creates. This is used in the new test `TestMultiStoreEventFeed`, which
passes a feed to multiple stores and verifies that the correct events are
published to the feed.

Finally, this commit renames "NewStoreEvent" to "AddStoreEvent".