Publish UpdateStoreEvent after Write commands
Ranges now publish an UpdateStoreEvent to their Store's event feed whenever a
write command completes. This event includes the command type that was executed
and the difference in stats that resulted from executing the command.

Included is a set of benchmarks which demonstrate that the generation of Update
events currently has a negligible impact on the performance of write commands.
This is accomplished by benchmarking write command with and without an event
feed.