Machinery to resolve write intents during the course of inconsistent reads.
Previously, inconsistent reads would simply ignore write intents. This is
not correct for two reasons. The biggest is that we do range lookups
with inconsistent reads. If a dangling intent is left on a meta record,
this means that we'd never have a way to get to the up-to-date meta
record, which might contain the latest information about a split, merge
or replica change. The second, smaller reason is that keys which are
only read inconsistently (say for performance) would have to wait for
a GC event in order to be resolved, which could be a long time.

This change does two things:

- Changes MVCC to continue a scan in the face of write intent errors
  and batch up all of them into a list of intents, to be resolved in
  parallel. Inconsistent reads continue, as before, in spite of intents,
  but now return a write intent error containing all intents encountered.
- At the store level, we now formulate an internal batch command to
  do this efficiently.

There's a big caveat which is documented in the code related to infinte
loops caused by intents encountered on range meta records. The basic
gist is that the intent is encountered on range lookup, which then
causes the store to push the txn, which causes the distributed sender
to lookup the txn. Since the txn record is always located on the range
which the meta record with the intent points to, this would in turn
cause another write intent error, and the cycle would repeat. The
solution to this is hairy and makes me nervous, but I couldn't find
any less obnoxious way.