move certs mock to *_test
this was needed in `security` to make `errcheck` happy,
and it is probably better to do it everywhere.

`make errcheck` still finds a lot of errors, but
now we can actually run it again. previously it
would try to find a package `security_test` after
complaining that `ResetTest()` (from main_test.go)
was undefined.