attempt at fixing #893
when a group gets deleted, everyone still waiting for a command to commit will
have their proposals aborted.
closes #893

there are still issues with this test (or multiraft/raft, depending on how you
see it), but this should fix the overwhelming majority of test failures in
TestStoreRangeMergeNonConsecutive.

see #906 for a very rare, open bug caused by this test.