disallow the log.V(i).Infof syntax
if you use it, you'll call Infof directly from the glog package,
and not from our local log package. That hasn't been an issue
so far, but it will be soon so this puts an end to it.

We might have an API like that again later, but let's keep it
simple for now.