For discussion: Introduce client.Callable interface.
This lets us use separate types for each operation (I've implemented one
for Get in this change), giving us type-safe accessors and a place to
potentially add helper methods.

This helps eliminate type assertions in application code, although it
is intrusive in the case of batch operations. I think this could be
alleviated with a new Calls type instead of a raw slice, but is
it worth it given the new higher-level client interface? (One reason
I'm interested in enhancing the Call-based interface is that I want to
use Calls in storage-level tests, where we can't always use the full
client)