raft rapid group recreation test
this simulates a rare error surfaced by #906 (and apparently others).
the test has a handful of concurrent actors add, commit to
(with retries) and remove a group.

currently this sometimes panics Raft:

E0506 13:09:45.552958   78550 log.go:188] applied(304) is out of range [prevApplied(10), committed(303)]
panic: applied(304) is out of range [prevApplied(10), committed(303)]

and other times (especially with a high number of writers) there are
deadlock issues.

i will continue to investigate these phenomena and hopefully attach
a fix to this PR once it merges.