Time Series now store Float64 exclusively
When Time Series storage was originally developed, we allowed time series data
to be expressed as either Integer or Float values.

However, while building the Query portion of the /ts package and revisiting this
two-type design, I now believe that there is insufficient reason to store
integer data. This commit removes the integer pathways, which is a significant
simplification.

This commmit also increases the size of floats stored; double-precision floats
are now used for all time series values.