Added acceptance directory and framework for acceptance tests.
The acceptance/localcluster package contains tools for starting up a
local cockroach cluster using docker. The acceptance package contains a
small framework for registering and running acceptance tests.

Ported the existing "gossip-peerings" acceptance test from
run/local-cluster.sh to the new acceptance framework.

Added a new build/builder.sh script which uses the cockroachdb/builder
image and various volume mappings to build linux binaries from the local
source tree.

Added an "install" target to the Makefile which is similar to "build"
but uses "go install" instead. The benefit of "install" target is that
it will not rebuild the output if the sources have not changed.