Tweak TestStoreRangeIterator and avoid creating overlapping ranges
The test used to create range ["a00", "a01") and ["a000", "a001").
This commit changes the test to split ["a00", "a01") to ["a00", "a000") and ["a000", "a01").

This fix was needed for me to pass the test with btree ranges, but I'm
not sure this change is right.. TestStoreAddRemoveRanges has a test
case for duplicated ranges, but I wasn't sure whether such range
overlapping can happen in real (and if so, how it should be
handled). If we need to support duplicated/overlapping ranges, I need to think
more on how to use btree.