fix race #967
this looks more like a problem with the test: the test is
waiting for a snapshot to be applied but checks this by
polling the engine, which means that it will go ahead
once the batch which applies the snapshot commits.

since the range stats were updated only after that,
there was a slight chance for a data race.

i have moved up the update to the stats object when
applying the snapshot since it did not seem useful
to switch to general atomic updates.