fix #1010
the issue was that coalesced heartbeats addressed their messages
to a dummy group. however, unreachable information was still
passed down to multinode, which in turn caused a message to
that dummy group to be passed into raft in those tests
which would stop a multinode as part of their tasks.
raft does not like receiving messages for groups it has no knowledge
of, so this caused a panic.
it's kind of scary that all of this time we've basically never had
working coalesced heartbeats despite the tests before today.
also, running test clusters must have had their leadership flipflop
all the time without us noticing. we need to bump up the logging for
coalesced heartbeat timeouts; those things should be visible.